requirements.txt
fastapi==0.95.1
uvicorn==0.18.3
authlib==1.2.0
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.25.0
httpx==0.24.0
itsdangerous==2.1.2
